backend/requirements_rpi.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
python-multipart==0.0.6
RPi.GPIO==0.7.0
adafruit-circuitpython-dht==3.7.10
adafruit-circuitpython-ads1x15==1.10.3
gpiozero==2.0.1
numpy==1.24.3
pandas==2.0.3
python-dateutil==2.8.2
picamera2